Added checkboxes and styling for todo list display
We've given users the ability to check whether or not a todo item has
been completed. In doing so the parent list will update the isCompleted
flag. Strike-through styling has also been added in for when an item has
been completed.

However, the main call-out is that when a new to-do item has been added,
styling does not persist however the parent list does maintaining the
correct flag for isCompleted.
	modified:   dist/main.js
	modified:   dist/main.js.map
	modified:   package-lock.json
	modified:   package.json
	modified:   src/index.js
	modified:   src/listDisplay.service.js
	modified:   src/listInterface.service.js
	modified:   src/newItem.service.js
	new file:   src/style.css
	modified:   webpack.config.js