Updated list display to automatically reorganize
The todo list display now automatically reorganizes to push completed
items to the bottom half of the list and items to-be completed at the
top.

The calendar input in the form was also updated to ensure the default
due date was always in place as it only showed for the first item being
added after page load.

 Changes to be committed:
	modified:   src/index.js
	modified:   src/listInterface.service.js